+++ libhfs/os/unix.c
@@ -19,6 +19,8 @@
  * $Id: unix.c,v 1.8 1998/11/02 22:09:13 rob Exp $
  */
 
+#include <sys/fcntl.h>
+
 #ifdef __linux__
 #define _FILE_OFFSET_BITS 64
 #define _LARGE_FILES
@@ -46,6 +48,7 @@ int stat(const char *, struct stat *);
 int fstat(int, struct stat *);
 # endif
 
+#  include <stdio.h>
 # include <errno.h>
 # include <sys/stat.h>
 
@@ -88,7 +91,7 @@ int os_open(void **priv, const char *path, int mode)
       (errno == EACCES || errno == EAGAIN))
     ERROR(EAGAIN, "unable to obtain lock for medium");
 
-  *priv = (void *) fd;
+  *priv = (void *) (long)fd;
 
   return 0;
 
@@ -105,7 +108,7 @@ fail:
  */
 int os_close(void **priv)
 {
-  int fd = (int) *priv;
+  int fd = (int)(long)*priv;
 
   *priv = (void *) -1;
 
@@ -124,7 +127,7 @@ fail:
  */
 int os_same(void **priv, const char *path)
 {
-  int fd = (int) *priv;
+  int fd = (int)(long)*priv;
   struct stat fdev, dev;
 
   if (fstat(fd, &fdev) == -1 ||
@@ -144,7 +147,7 @@ fail:
  */
 unsigned long os_seek(void **priv, unsigned long offset)
 {
-  int fd = (int) *priv;
+  int fd = (int)(long) *priv;
   off_t result;
 
   /* offset == -1 special; seek to last block of device */
@@ -169,7 +172,7 @@ fail:
  */
 unsigned long os_read(void **priv, void *buf, unsigned long len)
 {
-  int fd = (int) *priv;
+  int fd = (int)(long) *priv;
   ssize_t result;
 
   result = read(fd, buf, len << HFS_BLOCKSZ_BITS);
@@ -189,7 +192,7 @@ fail:
  */
 unsigned long os_write(void **priv, const void *buf, unsigned long len)
 {
-  int fd = (int) *priv;
+  int fd = (int)(long) *priv;
   ssize_t result;
 
   result = write(fd, buf, len << HFS_BLOCKSZ_BITS);
